LandSea

Documentation for LandSea.

LandSea.LandSeaData

LandSea.LandSeaFlat

LandSea.LandSeaTopo

LandSea.getLandSea

LandSea.LandSeaData — Type

LandSeaData

Abstract supertype for LandSea Datasets. All LandSeaData types contain the following fields:

lon - Vector containing the longitude points for the Land-Sea Dataset

lat - Vector containing the latitude points for the Land-Sea Dataset

lsm - Array containing data regarding the Land-Sea Mask. 1 is Land, 0 is Ocean, NaN is outside the bounds of the GeoRegion

LandSea.LandSeaFlat — Type

LandSeaFlat <: LandSeaData

A LandSea Dataset that contains only information on the land-sea mask and no topography.

LandSea.LandSeaTopo — Type

LandSeaTopo <: LandSeaData

A LandSea Dataset that also contains information on the topographic height.

A LandSeaTopo type will also contain the following field:

z - Array containing data regarding the Orographic Height in meters. NaN is outside the bounds of the GeoRegion

LandSea.getLandSea — Function

getLandSea

An extensible function type to retrieve LandSea Datasets. You can use this function name in your packages if you want to retrieve a specific LandSea dataset.